+++ applications/fishing-map/src/features/map/Map.tsx
@@ -1,6 +1,5 @@
 import React, { useCallback, useState, useEffect, useMemo } from 'react'
 import { createPortal } from 'react-dom'
-import { scaleLinear } from 'd3-scale'
 import { useSelector, useDispatch } from 'react-redux'
 import { useTranslation } from 'react-i18next'
 import { MapLegend, Tooltip } from '@globalfishingwatch/ui-components/dist'
@@ -21,7 +20,7 @@ import useMapInstance from 'features/map/map-context.hooks'
 import i18n from 'features/i18n/i18n'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
 import { useClickedEventConnect, useMapTooltip, useGeneratorsConnect } from 'features/map/map.hooks'
-import { selectDataviewInstancesResolved } from 'features/workspace/workspace.selectors'
+import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.selectors'
 import { selectEditing, moveCurrentRuler } from 'features/map/controls/rulers.slice'
 import MapInfo from 'features/map/controls/MapInfo'
 import MapControls from 'features/map/controls/MapControls'
@@ -35,8 +34,6 @@ import { SliceInteractionEvent } from './map.slice'
 import { useHasSourceLoaded } from './map-features.hooks'
 
 import '@globalfishingwatch/mapbox-gl/dist/mapbox-gl.css'
-
-const clickRadiusScale = scaleLinear().domain([4, 12, 17]).rangeRound([1, 2, 8]).clamp(true)
 
 // TODO: Abstract this away
 const transformRequest: (...args: any[]) => MapRequest = (url: string, resourceType: string) => {
@@ -210,7 +207,6 @@ const MapWrapper = (): React.ReactElement | null => {
           onResize={setMapBounds}
           getCursor={rulersEditing ? getRulersCursor : getCursor}
           interactiveLayerIds={rulersEditing ? undefined : style?.metadata?.interactiveLayerIds}
-          clickRadius={clickRadiusScale(viewport.zoom)}
           onClick={onMapClick}
           onHover={onMapHover}
           onError={handleError}