+++ applications/vessel-history/src/store.ts
@@ -7,6 +7,7 @@ import {
 } from '@reduxjs/toolkit'
 import connectedRoutes, { routerQueryMiddleware } from 'routes/routes'
 import vesselsReducer from 'features/vessels/vessels.slice'
+import searchReducer from 'features/search/search.slice'
 import mapReducer from './features/map/map.slice'
 
 const {
@@ -17,6 +18,7 @@ const {
 
 const rootReducer = combineReducers({
   vessels: vesselsReducer,
+  search: searchReducer,
   location: location,
   map: mapReducer,
 })