+++ packages/ui-components/src/map-legend/ColorRamp.tsx
@@ -23,15 +23,21 @@ function ColorRampLegend({
 
   // Omit bucket that goes from -Infinity --> 0. Will have to add an exception if we need a divergent scale
   const cleanRamp = ramp?.filter(([value]) => value !== Number.NEGATIVE_INFINITY)
-
-  const skipOddLabels = ramp && ramp.length > 6
+  const cleanValues = ramp?.filter(([value]) => value)
+  const skipOddLabels = cleanValues && cleanValues.length >= 6
 
   const heatmapLegendScale = useMemo(() => {
     if (!ramp || !cleanRamp) return null
 
+    const domainValues = cleanRamp.map(([value]) => {
+      if (value === 'less') return 0
+      if (value === 'more') return 1
+      return value as number
+    })
+
     return scaleLinear()
       .range(cleanRamp.map((item, i) => (i * 100) / (ramp.length - 1)))
-      .domain(cleanRamp.map(([value]) => value as number))
+      .domain(domainValues)
   }, [cleanRamp, ramp])
 
   if (!ramp || !cleanRamp) return null
@@ -91,11 +97,13 @@ function ColorRampLegend({
               const roundValue = roundValues
                 ? roundLegendNumber(value as number)
                 : (value as number)
-              const valueLabel = formatLegendValue(roundValue)
+              const valueLabel = typeof value === 'string' ? value : formatLegendValue(roundValue)
               if (skipOddLabels && i !== 0 && i !== ramp.length && i % 2 === 1) return null
               return (
                 <span
-                  className={styles.step}
+                  className={cx(styles.step, {
+                    [styles.lastStep]: !skipOddLabels && i === cleanRamp.length - 1,
+                  })}
                   style={{ left: `${(i * 100) / (ramp.length - 1)}%` }}
                   key={i}
                 >