+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.25
+
+### Patch Changes
+
+- Updated dependencies [bf2d09a1]
+  - @globalfishingwatch/layer-composer@5.2.0
+  - @globalfishingwatch/react-hooks@6.0.5
+  - @globalfishingwatch/ui-components@2.18.5
+
+## 0.28.24
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@6.2.2
+
+## 0.28.23
+
+### Patch Changes
+
+- Updated dependencies [af994916]
+  - @globalfishingwatch/layer-composer@5.1.2
+  - @globalfishingwatch/react-hooks@6.0.4
+  - @globalfishingwatch/ui-components@2.18.4
+
+## 0.28.22
+
+### Patch Changes
+
+- Updated dependencies [80907c1f]
+  - @globalfishingwatch/dataviews-client@6.2.1
+  - @globalfishingwatch/layer-composer@5.1.1
+  - @globalfishingwatch/react-hooks@6.0.3
+  - @globalfishingwatch/ui-components@2.18.3
+
+## 0.28.21
+
+### Patch Changes
+
+- Updated dependencies [87a54e62]
+  - @globalfishingwatch/dataviews-client@6.2.0
+  - @globalfishingwatch/layer-composer@5.1.0
+  - @globalfishingwatch/react-hooks@6.0.2
+  - @globalfishingwatch/ui-components@2.18.2
+
+## 0.28.20
+
+### Patch Changes
+
+- Updated dependencies [bb4978cd]
+  - @globalfishingwatch/layer-composer@5.0.1
+  - @globalfishingwatch/ui-components@2.18.1
+  - @globalfishingwatch/react-hooks@6.0.1
+
+## 0.28.19
+
+### Patch Changes
+
+- Updated dependencies [fdad03d6]
+  - @globalfishingwatch/ui-components@2.18.0
+
 ## 0.28.18
 
 ### Patch Changes